Correct machine type to resolve HVF memory errors
Reverts the previous addition of `tcg-vmsplit=1` and instead removes
the `highmem=off` parameter from the default machine type.

The previous commit attempted to fix a 32-bit memory addressing error
by adding `tcg-vmsplit=1`. This was incorrect, as it led to a new
error indicating the property was not found:
`Property 'virt-10.1-machine.tcg-vmsplit' not found`

This is because `tcg-vmsplit` is an option for QEMU's TCG accelerator,
not for the Hypervisor Framework (HVF) used on macOS.

The root cause of the original memory error was the `highmem=off`
parameter itself, which enforces a 32-bit guest address space. With
modern versions of QEMU on Apple Silicon, this flag is no longer
necessary and conflicts with memory allocations of 4GB or more.

By removing `highmem=off` and setting the machine type to simply `virt`,
we resolve the memory addressing limitations for HVF, allowing the VM to
boot successfully with the intended amount of RAM.