boot.py: Encapsulate isoinfo parsing logic
The logic for parsing `isoinfo` output was spread across several
loosely-related helper functions, making it difficult to follow.

To improve clarity and organization, this commit introduces a new
internal class, `_IsoInfoParser`. This class encapsulates all the
functionality required to parse the output of the `isoinfo` command
to find kernel and initrd candidates.

The previous helper functions have been converted into methods of this
new class, and `_find_kernel_initrd_with_isoinfo` is now simplified
to use an instance of the parser. This change improves modularity
without altering behavior.